sync fork

+++ conans/client/conf/config_installer.py
@@ -34,9 +34,10 @@ def _process_git_repo(repo_url, client_cache, output, runner, tmp_folder):
     _process_folder(tmp_folder, client_cache, output)
 
 
-def _process_zip_file(zippath, client_cache, output, tmp_folder):
+def _process_zip_file(zippath, client_cache, output, tmp_folder, remove=False):
     unzip(zippath, tmp_folder)
-    os.unlink(zippath)
+    if remove:
+        os.unlink(zippath)
     _process_folder(tmp_folder, client_cache, output)
 
 
@@ -62,8 +63,11 @@ def _process_folder(folder, client_cache, output):
                 registry_path = client_cache.registry
                 _handle_remotes(registry_path, os.path.join(root, f), output)
             else:
-                output.info("Copying file %s to %s" % (f, client_cache.conan_folder))
-                shutil.copy(os.path.join(root, f), client_cache.conan_folder)
+                relpath = os.path.relpath(root, folder)
+                target_folder = os.path.join(client_cache.conan_folder, relpath)
+                mkdir(target_folder)
+                output.info("Copying file %s to %s" % (f, target_folder))
+                shutil.copy(os.path.join(root, f), target_folder)
         for d in dirs:
             if d == "profiles":
                 output.info("Installing profiles")
@@ -77,7 +81,7 @@ def _process_download(item, client_cache, output, tmp_folder):
     output.info("Trying to download  %s" % item)
     zippath = os.path.join(tmp_folder, "config.zip")
     tools.download(item, zippath, out=output)
-    _process_zip_file(zippath, client_cache, output, tmp_folder)
+    _process_zip_file(zippath, client_cache, output, tmp_folder, remove=True)
 
 
 def configuration_install(item, client_cache, output, runner):
